Reduced movement speed granted by chests to balance out with the change to FixedUpdate-based movement.

+++ Assets/Scripts/Game Scripts/PlayerMovement.cs
@@ -92,7 +92,7 @@ public class PlayerMovement : MonoBehaviour {
             PlayerCombat.instance.StartDeath();
         }
     }
-    public void BuffMovement(int amount)
+    public void BuffMovement(float amount)
     {
         playerSpeed += amount;
     }
